refactor(ui): make the Serial tab's log and the monitor the same widget
#78 added a detachable serial monitor with autoscroll, timestamps, pause,
a filter, RX/TX colouring, Save… and a line-ending selector, and left the
Serial Config tab's "Serial monitor / debug" panel as it was: one colour,
a 500-line cap, a Command box that went through `send_command`, and no way
to narrow anything down. Two logs for the same traffic, only one of them
worth reading.

Extract everything below the monitor's connection header into
`ui/serial_console.py` and embed it in both places, so a feature added to
one is in the other by construction rather than by remembering. Two
constructor options mark the genuine differences: `show_baud` (the window's
speed picker — the tab's Connection panel already owns that setting, and two
controls writing one key is how they disagree) and `detach_command` (the
tab's "Open monitor ↗", on the control row so it costs no extra height).

The console subscribes to `serial/rx|tx|note` itself, so `app._log_serial`
no longer fans out to the tab and is just the backlog it was always feeding;
`SerialTab.append_log` and its `Text` are gone. `set_theme` repaints every
live console's Text tags, not only an open window's.

Tests follow the split: the behaviour cases move to
`tests/unit/ui/test_serial_console.py`, `test_serial_monitor.py` keeps what
the window adds (header, state, lifecycle), and `test_tab_serial.py` pins
that the tab's panel really is the shared console.
