+++ src/sorter/ui/tab_serial.py
@@ -12,6 +12,7 @@ from tkinter import messagebox, ttk
 from ..control.events import EventBus
 from ..hardware import serial_broker
 from ..hardware.serial_emulator import EMULATED_PORT
+from .serial_console import SerialConsole
 from .widgets import NumericField, build_button_row
 
 # (UI label, init-settings key, min, max, default). Defaults are the
@@ -168,23 +169,17 @@ class SerialTab(ttk.Frame):
             self.init_widgets[key] = field
 
         # ---- monitor & debug ----
+        # The same widget the detached window uses, minus its baud selector —
+        # the Connection panel above already owns that setting.
         monitor = ttk.LabelFrame(self, text="Serial monitor / debug")
         monitor.pack(side=tk.TOP, fill=tk.BOTH, expand=True, padx=8, pady=8)
-
-        cmd_row = ttk.Frame(monitor)
-        cmd_row.pack(side=tk.TOP, fill=tk.X, padx=4, pady=4)
-        ttk.Label(cmd_row, text="Command").pack(side=tk.LEFT, padx=4)
-        self.cmd_var = tk.StringVar()
-        entry = ttk.Entry(cmd_row, textvariable=self.cmd_var, width=40)
-        entry.pack(side=tk.LEFT, padx=4)
-        entry.bind("<Return>", lambda _e: self.send_command())
-        ttk.Button(cmd_row, text="Send", command=self.send_command).pack(side=tk.LEFT)
-        ttk.Button(cmd_row, text="Open monitor ↗", command=self.app.open_serial_monitor).pack(
-            side=tk.LEFT, padx=(12, 0)
+        self.console = SerialConsole(
+            monitor,
+            app=self.app,
+            height=10,
+            detach_command=self.app.open_serial_monitor,
         )
-
-        self.log = tk.Text(monitor, height=10, wrap=tk.NONE, state=tk.DISABLED)
-        self.log.pack(side=tk.TOP, fill=tk.BOTH, expand=True, padx=4, pady=4)
+        self.console.pack(side=tk.TOP, fill=tk.BOTH, expand=True, padx=6, pady=6)
 
         self.refresh_ports()
 
@@ -298,26 +293,3 @@ class SerialTab(ttk.Frame):
                 except (TypeError, ValueError, tk.TclError):
                     pass
         self.app.set_status(f"Loaded {applied} value(s) from board.")
-
-    def send_command(self) -> None:
-        broker = self.app.broker
-        if broker is None:
-            messagebox.showerror("Not connected", "Connect to the board first.")
-            return
-        cmd = self.cmd_var.get().strip()
-        if not cmd:
-            return
-        broker.send_command(cmd)
-        self.cmd_var.set("")
-
-    # ----- log API used by the App --------------------------------------------
-
-    def append_log(self, line: str) -> None:
-        self.log.configure(state=tk.NORMAL)
-        self.log.insert(tk.END, line + "\n")
-        self.log.see(tk.END)
-        # Trim to last ~500 lines.
-        line_count = int(self.log.index("end-1c").split(".")[0])
-        if line_count > 600:
-            self.log.delete("1.0", f"{line_count - 500}.0")
-        self.log.configure(state=tk.DISABLED)
